requirements.txt
flask>=2.0.1
requests>=2.26.0
python-dotenv>=0.19.0
openai>=1.0.0
youtube-transcript-api>=0.4.4
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.0.0
beautifulsoup4>=4.9.3
streamlit>=1.0.0
plotly>=5.0.0
